Rename ct_choice module to const_choice
